training/requirements-train.txt
# ============================================
# YOLO12 四部和聲 OMR 訓練環境依賴
# 硬體需求: RTX 5060 (8GB VRAM)
# Python 版本: 3.10+
# ============================================

# 深度學習框架
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Ultralytics YOLO12 (2025-02 最新版)
ultralytics>=8.3.0

# 資料處理
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0

# 視覺化
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# 進度條與工具
tqdm>=4.65.0
PyYAML>=6.0
scikit-learn>=1.3.0

# TFLite 轉換
tensorflow>=2.14.0
onnx>=1.15.0
onnxruntime>=1.16.0

# OMR 資料集處理
lxml>=4.9.0  # XML parsing for MUSCIMA++
h5py>=3.9.0  # HDF5 support

# 其他工具
requests>=2.31.0
